+++ src/lib/serialization/schema.ts
@@ -26,7 +26,6 @@
  *
  * For documentation on the JSON output properties, view the corresponding model.
  * @module
- * @internal
  */
 
 import type * as M from "../models";
@@ -271,6 +270,7 @@ export interface ReferenceType
         > {
     target: number | ReflectionSymbolId;
     qualifiedName?: string;
+    refersToTypeParameter?: boolean;
 }
 
 export interface ReflectionType